Restructure the TestWebService Expose different Apodini Libraries as products
