+++ Sources/TestWebService/UserSubsystem/Models/User.swift
@@ -1,0 +1,13 @@
+//
+//  User.swift
+//  
+//
+//  Created by Paul Schmiedmayer on 1/19/21.
+//
+
+import Apodini
+
+
+struct User: Codable, ResponseTransformable {
+    var id: Int
+}
